requirements.txt
autopep8==1.5.6
numpy==1.20.1
opencv-contrib-python==4.1.0.25
picamera==1.13
pkg-resources==0.0.0
pycodestyle==2.7.0
RPi.GPIO==0.7.0
toml==0.10.2
